+++ example/src/app.tsx
@@ -1,0 +1,11 @@
+import * as React from "react";
+
+import { Welcome } from "./pages/welcome";
+import { GlobalStyle } from "./app.styles";
+
+export const App = () => (
+  <>
+    <GlobalStyle />
+    <Welcome />
+  </>
+);