+++ src/state/index.ts
@@ -1,5 +1,6 @@
-export { ThemeState, ThemeDispatch } from "./theme-state";
 export * from "./actions";
+export * from "./context";
 export * from "./hook";
 export * from "./provider";
 export * from "./reducer";
+export { ThemeDispatch, ThemeState } from "./theme-state";